backEnd/vision_layer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision_layer"
version = "0.1.0"
description = "Detection-first vision processing for Basis cost segregation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Basis Team"}
]

dependencies = [
    # API clients
    "replicate>=0.25.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",

    # Image processing
    "pillow>=10.0.0",
    "numpy>=1.24.0",

    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Firebase
    "firebase-admin>=6.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "mcp_tools*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
